requirements.txt
Pillow==10.2.0
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3
matplotlib==3.7.1
tqdm==4.65.0
scikit-learn==1.3.0
seaborn==0.12.2 